+++ app/Modules/License/Database/Migrations/2026_04_08_000010_create_installations_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
     {
         Schema::create('installations', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('product_id');
-            $table->unsignedBigInteger('user_id');
+            $table->unsignedInteger('product_id');
+            $table->unsignedInteger('user_id');
             $table->string('license_code')->index();
             $table->string('installation_ip')->nullable();
             $table->string('installation_domain')->nullable();
