+++ app/Modules/License/Database/Migrations/2026_04_08_000014_create_version_callbacks_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
     {
         Schema::create('version_callbacks', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('product_id');
+            $table->unsignedInteger('product_id');
             $table->unsignedBigInteger('version_id');
             $table->string('callback_type');
             $table->string('callback_ip')->nullable();
